# docs/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autograder API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sandbox/courses/{course_id}/assignments/{assignment_id}/runs": {
      "post": {
        "tags": [
          "sandbox"
        ],
        "summary": "Create Run",
        "description": "Create a new sandbox run for a submission bundle, with optional stdin input.",
        "operationId": "create_run_sandbox_courses__course_id__assignments__assignment_id__runs_post",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          },
          {
            "name": "x-sandbox-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sandbox-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_run_sandbox_courses__course_id__assignments__assignment_id__runs_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxRunCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/runs/{run_id}/status": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run Status",
        "operationId": "get_run_status_runs__run_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/staff/courses/{course_id}/assignments/{assignment_id}/runs": {
      "get": {
        "tags": [
          "staff-runs"
        ],
        "summary": "List Official Runs",
        "operationId": "list_official_runs_staff_courses__course_id__assignments__assignment_id__runs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummaryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/staff/courses/{course_id}/assignments/{assignment_id}/runs/{run_id}": {
      "get": {
        "tags": [
          "staff-runs"
        ],
        "summary": "Get Official Run",
        "operationId": "get_official_run_staff_courses__course_id__assignments__assignment_id__runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/staff/courses/{course_id}/assignments/{assignment_id}/runs/{run_id}/details": {
      "get": {
        "tags": [
          "staff-runs"
        ],
        "summary": "Get Official Run Details",
        "operationId": "get_official_run_details_staff_courses__course_id__assignments__assignment_id__runs__run_id__details_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/staff/courses/{course_id}/assignments/{assignment_id}/runs/{run_id}/export/csv": {
      "get": {
        "tags": [
          "staff-runs"
        ],
        "summary": "Export Official Run Csv",
        "operationId": "export_official_run_csv_staff_courses__course_id__assignments__assignment_id__runs__run_id__export_csv_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/staff/courses/{course_id}/assignments/{assignment_id}/runs/{run_id}/export/feedback": {
      "get": {
        "tags": [
          "staff-runs"
        ],
        "summary": "Export Official Run Feedback",
        "operationId": "export_official_run_feedback_staff_courses__course_id__assignments__assignment_id__runs__run_id__export_feedback_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/staff/courses/{course_id}/assignments/{assignment_id}/runs/{run_id}/cleanup": {
      "post": {
        "tags": [
          "staff-runs"
        ],
        "summary": "Cleanup Official Run",
        "operationId": "cleanup_official_run_staff_courses__course_id__assignments__assignment_id__runs__run_id__cleanup_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/staff/courses/{course_id}/assignments/{assignment_id}/runs/{run_id}/students/{canvas_id}/files": {
      "get": {
        "tags": [
          "staff-runs"
        ],
        "summary": "List Student Files",
        "operationId": "list_student_files_staff_courses__course_id__assignments__assignment_id__runs__run_id__students__canvas_id__files_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          },
          {
            "name": "canvas_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Canvas Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_run_sandbox_courses__course_id__assignments__assignment_id__runs_post": {
        "properties": {
          "bundle": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Bundle"
          },
          "stdin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stdin"
          }
        },
        "type": "object",
        "required": [
          "bundle"
        ],
        "title": "Body_create_run_sandbox_courses__course_id__assignments__assignment_id__runs_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunStatusResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "queue",
              "run",
              "complete",
              "failure"
            ],
            "title": "State"
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position"
          },
          "eta_band": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "under_1_min",
                  "1_to_3_min",
                  "3_to_5_min",
                  "over_5_min"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Band"
          },
          "counters": {
            "$ref": "#/components/schemas/RunCounters"
          },
          "backpressure": {
            "$ref": "#/components/schemas/QueueBackpressure"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "state",
          "counters",
          "backpressure"
        ],
        "title": "RunStatusResponse"
      },
...
      "SandboxRunCreateResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "sandbox_session": {
            "type": "string",
            "title": "Sandbox Session"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url"
          },
          "result_url": {
            "type": "string",
            "title": "Result Url"
          },
          "upload_quota": {
            "$ref": "#/components/schemas/UploadQuota"
          },
          "initial_status": {
            "$ref": "#/components/schemas/RunStatusResponse"
          },
          "file_preview": {
            "$ref": "#/components/schemas/FilePreviewMetadata"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "sandbox_session",
          "status_url",
          "result_url",
          "upload_quota",
          "initial_status",
          "file_preview"
        ],
        "title": "SandboxRunCreateResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}